+++ src/App.css
@@ -1,10 +1,3 @@
 .App {
   text-align: center;
 }
-
-header {
-  background-color: darkblue;
-  color:lightblue;
-  overflow:auto;
-  padding: 20px;
-}
